requirements.txt
guizero==1.5.0
pyserial==3.5
